Process across instruments

+++ model/__init__.py
@@ -7,11 +7,11 @@ from scipy.io import wavfile
 import numpy as np
 from multiprocessing import Pool
 from multiprocessing.dummy import Pool as ThreadPool
+from cache import save
 
-NOTES_DIR = "../Iowa Notes/Mono"
-SPIKES_DIR = "./spikes/mono"
-SPIKES_ABS_DIR = "./spikes/mono/abs"
-TONE_SPIKES_DIR = "./spikes/tone"
+NOTES_DIR = "../notes/Piano"
+SPIKES_DIR = "./spikes/piano/abs"
+RNG_SPIKES_DIR = "./spikes/piano/rng"
 
 MAX_VAL = 2**15
 
@@ -25,7 +25,7 @@ def normalize_to_range(array):
         return np.zeros_like(array, dtype=float)
     
     # Normalize the array to the range [-1, 1]
-    normalized_array = 2 * (array - min_val) / (max_val - min_val) - 1
+    normalized_array = (2 * ((array - min_val) / (max_val - min_val))) - 1
     return normalized_array
 
 def normalize_absolute(array):
@@ -41,9 +41,7 @@ def normalize_absolute(array):
     normalized_array = array / MAX_VAL
     return normalized_array
 
-def generate_spikes_sync(sound_data, duration=0.25):
-    np.random.seed(711)
-
+def generate_spikes_sync(sound_data, duration=0.25, seed=711):
     # get data
     input_signal_fs, input_signal = sound_data
     input_signal = input_signal[:int(duration*input_signal_fs)]
@@ -67,6 +65,7 @@ def generate_spikes_sync(sound_data, duration=0.25):
         spont_list=sponts,
         implnt=1, # 0 = approximate, 1 = actual Power Law
         num_spike_trains=1,
+        random_seed=seed,
 
         # What is returned
         return_vihcs=False,
@@ -77,23 +76,56 @@ def generate_spikes_sync(sound_data, duration=0.25):
     )
     return np.array(result["nervegram_spike_times"])
 
-def save_spikes(note):
-    sound_data = wavfile.read(os.path.join(NOTES_DIR, f"{note}.wav"))
+def save_spikes(note, instrument="piano"):
+    notes_dir = f"../notes/{instrument}"
+    spikes_dir = f"./spikes/{instrument}"
+
+    sound_data = wavfile.read(os.path.join(notes_dir, f"{note}.wav"))
     spike_times = generate_spikes_sync(sound_data)
     spike_times = spike_times[0]
     if spike_times.shape != (3500,18,100):
         print(f"ERROR: {note} has shape {spike_times.shape}")
 
-    path = os.path.join(SPIKES_ABS_DIR, f"{note}.npy")
-    # Extract the directory name from the file path
-    directory = os.path.dirname(path)
+    path = os.path.join(spikes_dir, f"{note}.npy")
+    save(path, spike_times)
+
+def first_n_primes(n):
+    primes = []
+    num = 2
+    while len(primes) < n:
+        is_prime = True
+        for i in range(2, int(num ** 0.5) + 1):
+            if num % i == 0:
+                is_prime = False
+                break
+        if is_prime:
+            primes.append(num)
+        num += 1
+    return primes
+
+def save_spikes_rng(note):
+    sound_data = wavfile.read(os.path.join(NOTES_DIR, f"{note}.wav"))
     
-    # Create the directory recursively if it doesn't exist
-    if directory and not os.path.exists(directory):
-        os.makedirs(directory)
-    
-    # Save the numpy array to the path
-    np.save(path, spike_times)
+    seeds = first_n_primes(30)
+    for seed in seeds:
+        if os.path.exists(os.path.join(RNG_SPIKES_DIR, f"{note}_{seed}.npy")):
+            continue
+        spike_times = generate_spikes_sync(sound_data, seed=seed)
+        spike_times = spike_times[0]
+        if spike_times.shape != (3500,18,100):
+            print(f"ERROR: {note} has shape {spike_times.shape}")
+
+        path = os.path.join(RNG_SPIKES_DIR, f"{note}_{seed}.npy")
+        # Extract the directory name from the file path
+        directory = os.path.dirname(path)
+        
+        # Create the directory recursively if it doesn't exist
+        if directory and not os.path.exists(directory):
+            os.makedirs(directory)
+        
+        # Save the numpy array to the path
+        np.save(path, spike_times)
+
 
 def sharp_to_flat(sharp_note):
     # Mapping of sharp notes to flat notes
@@ -109,20 +141,31 @@ def sharp_to_flat(sharp_note):
     # Return the flat equivalent if it exists, otherwise return the input note
     return sharp_to_flat_map.get(sharp_note, sharp_note)
 
-def get_spikes(note):
+def get_spikes(note, mode="regular", instrument="piano"):
     if note[1] == "#":
         note = sharp_to_flat(note[:2]) + note[2:]
-    return np.load(os.path.join(SPIKES_DIR, f"{note}.npy"))
+    
+    if mode == "regular":
+        return np.load(os.path.join(f"./spikes/{instrument}", f"{note}.npy"))
+    elif mode == "rng":
+        final = []
+        for filename in os.listdir(RNG_SPIKES_DIR):
+            if filename.startswith(note):
+                final.append(np.load(os.path.join(RNG_SPIKES_DIR, filename)))
+        return final
 
-def get_spikes_abs(note):
+def get_decoded_exp(note, normalize=True):
     if note[1] == "#":
         note = sharp_to_flat(note[:2]) + note[2:]
-    return np.load(os.path.join(SPIKES_ABS_DIR, f"{note}.npy"))
+    decoded = np.load(os.path.join("./decoded/exp", f"{note}.npy"))
+    if normalize:
+        decoded = normalize_to_range(decoded)
+    return decoded
 
 def generate_scale(scale):
     processed = [
         filename.split(".")[0] 
-            for filename in os.listdir(SPIKES_ABS_DIR)
+            for filename in os.listdir(SPIKES_DIR)
                 if filename.endswith(".npy")
     ]
 
@@ -139,49 +182,3 @@ def generate_scale(scale):
                 print(f"Already processed: {note}")
 
     print(f"\nDONE WITH SCALE {scale}\n")
-
-def generate_pure_tone(frequency, duration=0.25, sample_rate=44100):
-    """
-    Generate a pure tone of a given frequency and duration.
-
-    Parameters:
-    - frequency (float): The frequency of the tone in Hz.
-    - duration (float): The duration of the tone in seconds.
-    - sample_rate (int): The sample rate in Hz. Default is 44100.
-
-    Returns:
-    - numpy.ndarray: A NumPy array containing the samples of the pure tone.
-    """
-    # Create an array of time values
-    t = np.linspace(0, duration, int(sample_rate * duration), endpoint=False)
-
-    # Generate the samples using the sine function
-    samples = np.sin(2 * np.pi * frequency * t)
-    
-    return samples
-
-
-def save_spikes_tone(freq):
-    tone = generate_pure_tone(freq)
-    spike_times = generate_spikes_sync((44100, tone))
-    if spike_times.shape == (1, 3500,18,100):
-        spike_times = spike_times[0]
-    if spike_times.shape != (3500,18,100):
-        print(f"ERROR: {freq} tone spikes has shape {spike_times.shape}")
-    np.save(os.path.join(TONE_SPIKES_DIR, f"{freq}.npy"), spike_times)
-
-def get_tone_spikes(freq, attempts=0):
-    max_attempts = 3  # Maximum number of attempts
-    
-    try:
-        spikes = np.load(os.path.join(TONE_SPIKES_DIR, f"{freq}.npy"))
-        return spikes
-    
-    except OSError:
-        if attempts >= max_attempts:
-            print(f"Failed to get spikes for frequency {freq} after {max_attempts} attempts.")
-            return None
-        
-        save_spikes_tone(freq)
-        return get_tone_spikes(freq, attempts + 1)
-
